+++ ellipticcurve/publicKey.py
@@ -1,6 +1,6 @@
 from .math import Math
 from .point import Point
-from .curve import secp256k1, getCurveByOid
+from .curve import secp256k1, getByOid
 from .utils.pem import getPemContent, createPem
 from .utils.der import hexFromInt, parse, DerFieldType, encodeConstructed, encodePrimitive
 from .utils.binary import hexFromByteString, byteStringFromHex, intFromHex, base64FromByteString, byteStringFromBase64
@@ -50,7 +50,7 @@ class PublicKey:
                 ecdsaPublicKeyOid=_ecdsaPublicKeyOid,
                 actualOid=publicKeyOid,
             ))
-        curve = getCurveByOid(curveOid)
+        curve = getByOid(curveOid)
         return cls.fromString(string=pointString, curve=curve)
 
     @classmethod